telegram-bot/requirements.txt
python-telegram-bot[job-queue]==21.10
python-dotenv==1.1.0
openpyxl==3.1.5
pandas==2.2.3
aiosqlite==0.20.0
edge-tts>=6.1.0
python-docx==1.2.0
reportlab==4.4.10
fpdf2==2.8.6
